requirements.txt
streamlit>=1.28.0
ibm-watsonx-ai>=1.0.0
python-dotenv>=1.0.0
